requirements.txt
prompt_toolkit>=3.0.43
python-dotenv>=1.0.1
docker>=7.1.0
requests>=2.32.3
psutil>=7.2.1
